Introduce Valid_Flags (#77)
* Introduce Valid_Flags

Valid_options has a limit: only 1 can be selected. This doesn`t work when we use bitmask (like for errors) so we are resorting to having to write a giant description, which looks much, much worse than when the JSON can have clear separate items, each with a value and description.

Therefore, we introduce a new parameter type. In addition to Valid_Range (discrete range) and Valid_Options (distinct, single options), we introduce Valid_Flags

* Fix CO_PARAM_BMS_MISSING_CONFIG

* Update FTEX_Controller_CANOpen_Protocol.json

* Update FTEX_BMS_CANOpen_Protocol.json

* fix schema validator for new type

* fix PAS json

* Update FTEX_Controller_CANOpen_Protocol.json